Separate Calendar and Activity page, Adjust Welcome and Select Mood page layout
